requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-groq>=0.1.0
groq>=0.4.0
yfinance>=0.2.0
duckduckgo-search>=4.0.0
ddgs==9.5.4
python-dotenv>=1.0.0
pydantic>=2.0.0
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
markdown>=3.5.0
